Struct rusoto_iot1click_devices::Iot1ClickDevicesClient

pub struct Iot1ClickDevicesClient { /* fields omitted */ }

A client for the AWS IoT 1-Click Devices Service API.

Methods

impl Iot1ClickDevicesClient

pub fn new(region: Region) -> Iot1ClickDevicesClient

Creates a client backed by the default tokio event loop.

The client will use the default credentials provider and tls client.

pub fn new_with<P, D>(
request_dispatcher: D,
credentials_provider: P,
region: Region
) -> Iot1ClickDevicesClient where
P: ProvideAwsCredentials + Send + Sync + 'static,
P::Future: Send,
D: DispatchSignedRequest + Send + Sync + 'static,
D::Future: Send,