Trait rusoto_iot1click_devices::Iot1ClickDevices

pub trait Iot1ClickDevices {
    fn claim_devices_by_claim_code(
        &self, 
        input: ClaimDevicesByClaimCodeRequest
    ) -> RusotoFuture<ClaimDevicesByClaimCodeResponse, ClaimDevicesByClaimCodeError>;

    fn describe_device(
        &self, 
        input: DescribeDeviceRequest
    ) -> RusotoFuture<DescribeDeviceResponse, DescribeDeviceError>;

    fn finalize_device_claim(
        &self, 
        input: FinalizeDeviceClaimRequest
    ) -> RusotoFuture<FinalizeDeviceClaimResponse, FinalizeDeviceClaimError>;

    fn get_device_methods(
        &self, 
        input: GetDeviceMethodsRequest
    ) -> RusotoFuture<GetDeviceMethodsResponse, GetDeviceMethodsError>;

    fn initiate_device_claim(
        &self, 
        input: InitiateDeviceClaimRequest
    ) -> RusotoFuture<InitiateDeviceClaimResponse, InitiateDeviceClaimError>;

    fn invoke_device_method(
        &self, 
        input: InvokeDeviceMethodRequest
    ) -> RusotoFuture<InvokeDeviceMethodResponse, InvokeDeviceMethodError>;

    fn list_device_events(
        &self, 
        input: ListDeviceEventsRequest
    ) -> RusotoFuture<ListDeviceEventsResponse, ListDeviceEventsError>;

    fn list_devices(
        &self, 
        input: ListDevicesRequest
    ) -> RusotoFuture<ListDevicesResponse, ListDevicesError>;

    fn list_tags_for_resource(
        &self, 
        input: ListTagsForResourceRequest
    ) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>;

    fn tag_resource(
        &self, 
        input: TagResourceRequest
    ) -> RusotoFuture<(), TagResourceError>;

    fn unclaim_device(
        &self, 
        input: UnclaimDeviceRequest
    ) -> RusotoFuture<UnclaimDeviceResponse, UnclaimDeviceError>;

    fn untag_resource(
        &self, 
        input: UntagResourceRequest
    ) -> RusotoFuture<(), UntagResourceError>;

    fn update_device_state(
        &self, 
        input: UpdateDeviceStateRequest
    ) -> RusotoFuture<UpdateDeviceStateResponse, UpdateDeviceStateError>;
}

Trait representing the capabilities of the AWS IoT 1-Click Devices Service API. AWS IoT 1-Click Devices Service clients implement this trait.

Required methods

fn claim_devices_by_claim_code(
&self,
input: ClaimDevicesByClaimCodeRequest
) -> RusotoFuture<ClaimDevicesByClaimCodeResponse, ClaimDevicesByClaimCodeError>

Adds device(s) to your account (i.e., claim one or more devices) if and only if you received a claim code with the device(s).

fn describe_device(
&self,
input: DescribeDeviceRequest
) -> RusotoFuture<DescribeDeviceResponse, DescribeDeviceError>

Given a device ID, returns a DescribeDeviceResponse object describing the details of the device.

fn finalize_device_claim(
&self,
input: FinalizeDeviceClaimRequest
) -> RusotoFuture<FinalizeDeviceClaimResponse, FinalizeDeviceClaimError>

Given a device ID, finalizes the claim request for the associated device.

Claiming a device consists of initiating a claim, then publishing a device event, and finalizing the claim. For a device of type button, a device event can be published by simply clicking the device.

fn get_device_methods(
&self,
input: GetDeviceMethodsRequest
) -> RusotoFuture<GetDeviceMethodsResponse, GetDeviceMethodsError>

Given a device ID, returns the invokable methods associated with the device.

fn initiate_device_claim(
&self,
input: InitiateDeviceClaimRequest
) -> RusotoFuture<InitiateDeviceClaimResponse, InitiateDeviceClaimError>

Given a device ID, initiates a claim request for the associated device.

Claiming a device consists of initiating a claim, then publishing a device event, and finalizing the claim. For a device of type button, a device event can be published by simply clicking the device.

fn invoke_device_method(
&self,
input: InvokeDeviceMethodRequest
) -> RusotoFuture<InvokeDeviceMethodResponse, InvokeDeviceMethodError>

Given a device ID, issues a request to invoke a named device method (with possible parameters). See the "Example POST" code snippet below.

fn list_device_events(
&self,
input: ListDeviceEventsRequest
) -> RusotoFuture<ListDeviceEventsResponse, ListDeviceEventsError>

Using a device ID, returns a DeviceEventsResponse object containing an array of events for the device.

fn list_devices(
&self,
input: ListDevicesRequest
) -> RusotoFuture<ListDevicesResponse, ListDevicesError>

Lists the 1-Click compatible devices associated with your AWS account.

fn list_tags_for_resource(
&self,
input: ListTagsForResourceRequest
) -> RusotoFuture<ListTagsForResourceResponse, ListTagsForResourceError>

Lists the tags associated with the specified resource ARN.

fn tag_resource(
&self,
input: TagResourceRequest
) -> RusotoFuture<(), TagResourceError>

Adds or updates the tags associated with the resource ARN. See AWS IoT 1-Click Service Limits for the maximum number of tags allowed per resource.

fn unclaim_device(
&self,
input: UnclaimDeviceRequest
) -> RusotoFuture<UnclaimDeviceResponse, UnclaimDeviceError>

Disassociates a device from your AWS account using its device ID.

fn untag_resource(
&self,
input: UntagResourceRequest
) -> RusotoFuture<(), UntagResourceError>

Using tag keys, deletes the tags (key/value pairs) associated with the specified resource ARN.

fn update_device_state(
&self,
input: UpdateDeviceStateRequest
) -> RusotoFuture<UpdateDeviceStateResponse, UpdateDeviceStateError>

Using a Boolean value (true or false), this operation enables or disables the device given a device ID.

Loading content...